Reorganized AutoML packages for better package cohesions and single responsibility per package.
