+++ neuraxle/metaopt/auto_ml.py
@@ -36,16 +36,17 @@ from neuraxle.data_container import IDT
 from neuraxle.data_container import DataContainer as DACT
 from neuraxle.hyperparams.space import HyperparameterSpace
 from neuraxle.metaopt.callbacks import (ARG_Y_EXPECTED, ARG_Y_PREDICTD,
-                                        BaseCallback, CallbackList, MetricCallback,
-                                        ScoringCallback)
+                                        BaseCallback, CallbackList,
+                                        MetricCallback, ScoringCallback)
+from neuraxle.metaopt.context import AutoMLContext
 from neuraxle.metaopt.data.aggregates import (Client, Project, Root, Round,
                                               Trial, TrialSplit)
 from neuraxle.metaopt.data.reporting import RoundReport
 from neuraxle.metaopt.data.vanilla import (DEFAULT_CLIENT, DEFAULT_PROJECT,
-                                           AutoMLContext,
-                                           HyperparamsRepository, RoundDataclass, ScopedLocation)
-from neuraxle.metaopt.validation import (BaseHyperparameterOptimizer,
-                                         BaseValidationSplitter,
+                                           RoundDataclass, ScopedLocation)
+from neuraxle.metaopt.optimizer import BaseHyperparameterOptimizer
+from neuraxle.metaopt.repositories.repo import HyperparamsRepository
+from neuraxle.metaopt.validation import (BaseValidationSplitter,
                                          GridExplorationSampler,
                                          RandomSearchSampler,
                                          ValidationSplitter)
