+++ neuraxle/metaopt/context.py
@@ -1,0 +1,169 @@
+"""
+Neuraxle's AutoML Context Management
+====================================================
+
+The :class:`~neuraxle.base.ExecutionContext`, is a special object
+that needs to be adapted in AutoML loops.
+
+
+.. seealso::
+    :class:`~neuraxle.metaopt.hyperopt.tpe.TreeParzenEstimator`,
+
+
+..
+    Copyright 2022, Neuraxio Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+
+"""
+
+import logging
+
+from neuraxle.base import ContextLock
+from neuraxle.base import ExecutionContext as CX
+from neuraxle.base import NoContextLock
+from neuraxle.logging.logging import NeuraxleLogger
+from neuraxle.metaopt.data.vanilla import (BaseDataclass, ScopedLocation,
+                                           SubDataclassT)
+from neuraxle.metaopt.repositories.repo import (HyperparamsRepository,
+                                                VanillaHyperparamsRepository)
+
+
+class AutoMLContext(CX):
+
+    @property
+    def logger(self) -> NeuraxleLogger:
+        self.add_scoped_logger_file_handler()  # TODO: this is perhaps why logs are duplicated.
+        return CX.logger.fget(self)
+
+    @property
+    def logger_at_scoped_loc(self) -> NeuraxleLogger:
+        return logging.getLogger(self.get_identifier(include_step_names=False))
+
+    def add_scoped_logger_file_handler(self) -> NeuraxleLogger:
+        """
+        Add a file handler to the logger at the current scoped location to capture logs
+        at this scope and below this scope.
+        """
+        # TODO: with self.lock:
+        self.repo.add_logging_handler(self.logger_at_scoped_loc, self.loc)
+
+    def free_scoped_logger_file_handler(self):
+        """
+        Remove file handlers from logger to free file lock (especially on Windows).
+        """
+        self.logger_at_scoped_loc.without_file_handler()
+
+    def read_scoped_log(self) -> str:
+        """
+        Read the scoped logger file.
+        """
+        # TODO: with self.lock:
+        return self.repo.get_log_from_logging_handler(self.logger, self.loc)
+
+    def copy(self):
+        copy_kwargs = self._get_copy_kwargs()
+        return AutoMLContext(**copy_kwargs)
+
+    def _get_copy_kwargs(self):
+        return super()._get_copy_kwargs()
+
+    @staticmethod
+    def from_context(
+        context: CX = None,
+        repo: HyperparamsRepository = None,
+        loc: ScopedLocation = None,
+        disable_context_lock: bool = False,
+    ) -> 'AutoMLContext':
+        """
+        Create a new AutoMLContext from an ExecutionContext.
+
+        :param context: ExecutionContext
+        """
+        new_context: AutoMLContext = AutoMLContext.copy(
+            context if context is not None else AutoMLContext()
+        )
+        if disable_context_lock is True:
+            new_context.disable_context_lock()
+        if not new_context.has_service(HyperparamsRepository):
+            new_context.register_service(
+                HyperparamsRepository,
+                repo or VanillaHyperparamsRepository(new_context.get_path())
+            )
+        if not new_context.has_service(ScopedLocation):
+            new_context.register_service(
+                ScopedLocation,
+                loc or ScopedLocation()
+            )
+        if not new_context.has_service(ContextLock):
+            new_context.register_service(
+                ContextLock,
+                ContextLock()
+            )
+        return new_context
+
+    @property
+    def lock(self):
+        # TODO: Global locks are known to be something to get rid of, such as Python's GIL. Use service locks. Rm this.
+        return self.synchroneous()
+
+    def disable_context_lock(self) -> 'AutoMLContext':
+        self.register_service(ContextLock, NoContextLock())
+        return self
+
+    @property
+    def loc(self) -> ScopedLocation:
+        return self.get_service(ScopedLocation)
+
+    @property
+    def repo(self) -> HyperparamsRepository:
+        return self.get_service(HyperparamsRepository)
+
+    def push_attr(self, subdataclass: SubDataclassT) -> 'AutoMLContext':
+        """
+        Push a new attribute into the ScopedLocation.
+
+        :param name: attribute name
+        :param value: attribute value
+        :return: an AutoMLContext copy with the new loc attribute.
+        """
+        if not isinstance(subdataclass, BaseDataclass) and isinstance(subdataclass, (str, int)):
+            return self.with_loc(self.loc.with_id(subdataclass))  # ID instead.
+        return self.with_loc(self.loc.with_dc(subdataclass))
+
+    def pop_attr(self) -> 'AutoMLContext':
+        """
+        Pop an attribute from the ScopedLocation.
+
+        :return: an AutoMLContext copy with the new popped loc attribute.
+        """
+        return self.with_loc(self.loc.popped())
+
+    def with_loc(self, loc: ScopedLocation) -> 'AutoMLContext':
+        """
+        Replace the ScopedLocation by the one provided.
+
+        :param loc: ScopedLocation
+        :return: an AutoMLContext copy with the new loc attribute.
+        """
+        new_self: AutoMLContext = self.copy()
+        new_self.register_service(ScopedLocation, loc)
+        return new_self
+
+    def load_dc(self, deep=True) -> BaseDataclass:
+        """
+        Load the current dc from the repo.
+        """
+        with self.lock:
+            return self.repo.load(self.loc, deep)
