+++ neuraxle/metaopt/data/vanilla.py
@@ -30,12 +30,11 @@ Classes are splitted like this for the AutoML:
 
 
 """
+
 import copy
 import datetime
 import json
-import logging
 import math
-import os
 import typing
 from abc import ABC, abstractmethod
 from collections import OrderedDict
@@ -45,12 +44,10 @@ from typing import (Any, Dict, Generic, List, Optional, Sequence, Tuple, Type,
                     TypeVar, Union)
 
 import numpy as np
-from neuraxle.base import (CX, BaseService, ContextLock, NoContextLock,
-                           TrialStatus)
+from neuraxle.base import CX, BaseService, TrialStatus
 from neuraxle.hyperparams.space import HyperparameterSamples, RecursiveDict
-from neuraxle.logging.logging import (LOGGING_DATETIME_STR_FORMAT,
-                                      NeuraxleLogger)
-from neuraxle.logging.warnings import RaiseDeprecatedClass
+from neuraxle.logging.logging import LOGGING_DATETIME_STR_FORMAT
+
 
 SubDataclassT = TypeVar('SubDataclassT', bound=Optional['BaseDataclass'])
 ScopedLocationAttrInt = int
@@ -627,7 +624,8 @@ class DataclassHasListMixin:
         elif _id < self.get_next_i():
             self.get_sublocation()[dc.get_id()] = dc
         else:
-            raise ValueError(f"{dc.shallow()} has id {dc.get_id()} which is greater than the next id {self.get_next_i()}.")
+            raise ValueError(
+                f"{dc.shallow()} has id {dc.get_id()} which is greater than the next id {self.get_next_i()}.")
         self._validate()
         return _id
 
@@ -943,331 +941,3 @@ def to_json(obj: BaseDataclass) -> str:
 
 def from_json(_json: str) -> BaseDataclass:
     return json.loads(_json, object_pairs_hook=object_pairs_decoder, object_hook=object_decoder)
-
-
-class HyperparamsRepository(BaseService):
-    """
-    Hyperparams repository that saves hyperparams, and scores for every AutoML trial.
-    Cache folder can be changed to do different round numbers.
-
-    For more information, read this `article by Martin Fowler on DDD Aggregates <https://martinfowler.com/bliki/DDD_Aggregate.html>`_.
-
-    .. seealso::
-        :class:`AutoML`,
-        :class:`Trainer`,
-    """
-
-    def __init__(self):
-        BaseService.__init__(self)
-
-    @abstractmethod
-    def load(self, scope: ScopedLocation, deep=False) -> SubDataclassT:
-        """
-        Get metadata from scope.
-
-        The fetched metadata will be the one that is the last item
-        that is not a None in the provided scope.
-
-        :param scope: scope to get metadata from.
-        :return: metadata from scope.
-        """
-        raise NotImplementedError("Use a concrete class. This is an abstract class.")
-
-    @abstractmethod
-    def save(self, _dataclass: SubDataclassT, scope: ScopedLocation, deep=False) -> 'HyperparamsRepository':
-        """
-        Save metadata to scope.
-
-        :param metadata: metadata to save.
-        :param scope: scope to save metadata to.
-        :param deep: if True, save metadata's sublocations recursively.
-        """
-        raise NotImplementedError("Use a concrete class. This is an abstract class.")
-
-    @abstractmethod
-    def add_logging_handler(self, logger: NeuraxleLogger, scope: ScopedLocation) -> 'HyperparamsRepository':
-        """
-        Add logging handler to repository's provided scope.
-        Handler must be set manually for each scope below this scope.
-
-        :param logger: logger to add handler to.
-        :param scope: scope to add handler to.
-        :return: self.
-        """
-        raise NotImplementedError("Use a concrete class. This is an abstract class.")
-
-    @abstractmethod
-    def get_log_from_logging_handler(self, logger: NeuraxleLogger, scope: ScopedLocation) -> str:
-        """
-        Get log from repository's provided scope and handler that was set with :func:`add_logging_handler`.
-
-        :param scope: scope to get log from.
-        :return: log from scope.
-        """
-        raise NotImplementedError("Use a concrete class. This is an abstract class.")
-
-    @abstractmethod
-    def is_locking_required(self) -> bool:
-        """
-        Check if repository is locking required.
-        """
-        raise NotImplementedError("Use a concrete class. This is an abstract class.")
-
-
-class _InMemoryRepositoryLoggerHandlerMixin:
-    """
-    Mixin to add a in-memory logging handler to a repository.
-    """
-
-    def add_logging_handler(self, logger: NeuraxleLogger, scope: ScopedLocation) -> 'HyperparamsRepository':
-        return self
-
-    def get_log_from_logging_handler(self, logger: NeuraxleLogger, scope: ScopedLocation) -> str:
-        return logger.get_scoped_string_history()
-
-
-class VanillaHyperparamsRepository(_InMemoryRepositoryLoggerHandlerMixin, HyperparamsRepository):
-    """
-    Hyperparams repository that saves data AutoML-related info.
-    """
-
-    def __init__(
-        self,
-        cache_folder: str
-    ):
-        """
-        :param cache_folder: folder to store trials.
-        :param hyperparams_repo_class: class to use to save hyperparams.
-        :param hyperparams_repo_kwargs: kwargs to pass to hyperparams_repo_class.
-        """
-        HyperparamsRepository.__init__(self)
-        _InMemoryRepositoryLoggerHandlerMixin.__init__(self)
-        self.cache_folder = os.path.join(cache_folder, self.__class__.__name__)
-        self.root: RootDataclass = RootDataclass()
-
-    @staticmethod
-    def from_root(root: RootDataclass, cache_folder: str) -> 'VanillaHyperparamsRepository':
-
-        return VanillaHyperparamsRepository(
-            cache_folder=cache_folder,
-        ).save(root, ScopedLocation(), deep=True)
-
-    def load(self, scope: ScopedLocation, deep=False) -> SubDataclassT:
-        try:
-            ret: BaseDataclass = self.root[scope]
-            if not deep:
-                ret = ret.shallow()
-        except KeyError:
-            ret: BaseDataclass = scope.new_dataclass_from_id()
-
-        return copy.deepcopy(ret)
-
-    def save(self, _dataclass: SubDataclassT, scope: ScopedLocation, deep=False) -> 'VanillaHyperparamsRepository':
-        # Sanitizing
-        _dataclass: SubDataclassT = copy.deepcopy(_dataclass)
-        scope = scope.at_dc(_dataclass)
-
-        # Sanity checks: good type
-        if not isinstance(_dataclass, BaseDataclass):
-            raise ValueError(f"Was expecting a dataclass. Got `{_dataclass.__class__.__name__}`.")
-        # Sanity checks: sufficient scope depth
-        scope = scope[:_dataclass.__class__]  # Sanitizing scope to dtype loc.
-        _id_from_scope: ScopedLocationAttr = scope[_dataclass.__class__]
-        if _id_from_scope is not None:
-            if _id_from_scope != _dataclass.get_id():
-                raise ValueError(
-                    f"The scope `{scope}` with {_dataclass.__class__.__name__} id `{_id_from_scope}` does not match the provided dataclass id `{_dataclass.get_id()}` for `{_dataclass}`."
-                )
-            scope.pop()
-            # else check if the scope is at least of the good class length:
-        elif len(scope) != list(dataclass_2_subloc_attr.keys()).index(_dataclass.__class__):
-            raise ValueError(
-                f"The scope `{scope}` is not of the good length for dataclass type `{_dataclass.__class__.__name__}`."
-            )
-
-        # Passthrough dc's sublocation when saving shallow:
-        if not deep:
-            if isinstance(_dataclass, RootDataclass):
-                _dataclass.set_sublocation(self.root.get_sublocation())
-            elif scope.with_dc(_dataclass) in self.root:
-                prev_dc: SubDataclassT = self.root[scope.with_dc(_dataclass)]
-                _dataclass.set_sublocation(prev_dc.get_sublocation())
-            else:
-                _dataclass = _dataclass.empty()
-
-        # Finally storing.
-        if isinstance(_dataclass, RootDataclass):
-            self.root = _dataclass
-        else:
-            self.root[scope].store(_dataclass)
-        return self
-
-    def is_locking_required(self) -> bool:
-        return False
-
-
-class InMemoryHyperparamsRepository(RaiseDeprecatedClass, VanillaHyperparamsRepository):
-    """
-    In memory hyperparams repository that can print information about trials.
-    Useful for debugging.
-    """
-
-    def __init__(self, *kargs, **kwargs):
-        RaiseDeprecatedClass.__init__(
-            self,
-            replacement_class=VanillaHyperparamsRepository,
-            since_version="0.7.0",
-        )
-        VanillaHyperparamsRepository.__init__(self, *kargs, **kwargs)
-
-
-class BaseHyperparameterOptimizer(ABC):
-
-    @abstractmethod
-    def find_next_best_hyperparams(self, round_scope) -> HyperparameterSamples:
-        """
-        Find the next best hyperparams using previous trials, that is the
-        whole :class:`neuraxle.metaopt.data.aggregate.Round`.
-
-        :param round: a :class:`neuraxle.metaopt.data.aggregate.Round`
-        :return: next hyperparameter samples to train on
-        """
-        raise NotImplementedError()
-
-
-class HyperparameterSamplerStub(BaseHyperparameterOptimizer):
-
-    def __init__(self, preconfigured_hp_samples: HyperparameterSamples):
-        self.preconfigured_hp_samples = preconfigured_hp_samples
-
-    def find_next_best_hyperparams(self, round_scope) -> HyperparameterSamples:
-        return self.preconfigured_hp_samples
-
-
-class AutoMLContext(CX):
-
-    @property
-    def logger(self) -> NeuraxleLogger:
-        self.add_scoped_logger_file_handler()  # TODO: this is perhaps why logs are duplicated.
-        return CX.logger.fget(self)
-
-    @property
-    def logger_at_scoped_loc(self) -> NeuraxleLogger:
-        return logging.getLogger(self.get_identifier(include_step_names=False))
-
-    def add_scoped_logger_file_handler(self) -> NeuraxleLogger:
-        """
-        Add a file handler to the logger at the current scoped location to capture logs
-        at this scope and below this scope.
-        """
-        # TODO: with self.lock:
-        self.repo.add_logging_handler(self.logger_at_scoped_loc, self.loc)
-
-    def free_scoped_logger_file_handler(self):
-        """
-        Remove file handlers from logger to free file lock (especially on Windows).
-        """
-        self.logger_at_scoped_loc.without_file_handler()
-
-    def read_scoped_log(self) -> str:
-        """
-        Read the scoped logger file.
-        """
-        # TODO: with self.lock:
-        return self.repo.get_log_from_logging_handler(self.logger, self.loc)
-
-    def copy(self):
-        copy_kwargs = self._get_copy_kwargs()
-        return AutoMLContext(**copy_kwargs)
-
-    def _get_copy_kwargs(self):
-        return super()._get_copy_kwargs()
-
-    @staticmethod
-    def from_context(
-        context: CX = None,
-        repo: HyperparamsRepository = None,
-        loc: ScopedLocation = None,
-        disable_context_lock: bool = False,
-    ) -> 'AutoMLContext':
-        """
-        Create a new AutoMLContext from an ExecutionContext.
-
-        :param context: ExecutionContext
-        """
-        new_context: AutoMLContext = AutoMLContext.copy(
-            context if context is not None else AutoMLContext()
-        )
-        if disable_context_lock is True:
-            new_context.disable_context_lock()
-        if not new_context.has_service(HyperparamsRepository):
-            new_context.register_service(
-                HyperparamsRepository,
-                repo or VanillaHyperparamsRepository(new_context.get_path())
-            )
-        if not new_context.has_service(ScopedLocation):
-            new_context.register_service(
-                ScopedLocation,
-                loc or ScopedLocation()
-            )
-        if not new_context.has_service(ContextLock):
-            new_context.register_service(
-                ContextLock,
-                ContextLock()
-            )
-        return new_context
-
-    @property
-    def lock(self):
-        # TODO: Global locks are known to be something to get rid of, such as Python's GIL. Use service locks. Rm this.
-        return self.synchroneous()
-
-    def disable_context_lock(self) -> 'AutoMLContext':
-        self.register_service(ContextLock, NoContextLock())
-        return self
-
-    @property
-    def loc(self) -> ScopedLocation:
-        return self.get_service(ScopedLocation)
-
-    @property
-    def repo(self) -> HyperparamsRepository:
-        return self.get_service(HyperparamsRepository)
-
-    def push_attr(self, subdataclass: SubDataclassT) -> 'AutoMLContext':
-        """
-        Push a new attribute into the ScopedLocation.
-
-        :param name: attribute name
-        :param value: attribute value
-        :return: an AutoMLContext copy with the new loc attribute.
-        """
-        if not isinstance(subdataclass, BaseDataclass) and isinstance(subdataclass, (str, int)):
-            return self.with_loc(self.loc.with_id(subdataclass))  # ID instead.
-        return self.with_loc(self.loc.with_dc(subdataclass))
-
-    def pop_attr(self) -> 'AutoMLContext':
-        """
-        Pop an attribute from the ScopedLocation.
-
-        :return: an AutoMLContext copy with the new popped loc attribute.
-        """
-        return self.with_loc(self.loc.popped())
-
-    def with_loc(self, loc: ScopedLocation) -> 'AutoMLContext':
-        """
-        Replace the ScopedLocation by the one provided.
-
-        :param loc: ScopedLocation
-        :return: an AutoMLContext copy with the new loc attribute.
-        """
-        new_self: AutoMLContext = self.copy()
-        new_self.register_service(ScopedLocation, loc)
-        return new_self
-
-    def load_dc(self, deep=True) -> BaseDataclass:
-        """
-        Load the current dc from the repo.
-        """
-        with self.lock:
-            return self.repo.load(self.loc, deep)
