+++ testing_neuraxle/test_service_assertions.py
@@ -8,9 +8,9 @@ from neuraxle.base import (BaseService, BaseStep, BaseTransformerT,
                            CX, ForceHandleIdentity, Identity,
                            NonFittableMixin, StepWithContext)
 from neuraxle.data_container import DataContainer as DACT
-from neuraxle.metaopt.auto_ml import AutoML, RandomSearchSampler
+from neuraxle.metaopt.auto_ml import AutoML
 from neuraxle.metaopt.callbacks import ScoringCallback
-from neuraxle.metaopt.data.json_repo import HyperparamsOnDiskRepository
+from neuraxle.metaopt.repositories.json import HyperparamsOnDiskRepository
 from neuraxle.metaopt.validation import GridExplorationSampler, ValidationSplitter
 from neuraxle.pipeline import Pipeline
 from sklearn.metrics import mean_squared_error
